requirements.txt
flask==3.1.0
Pillow==11.1.0
supabase==2.13.0
python-dotenv==1.1.0
gunicorn==23.0.0
